requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
slack-sdk==3.21.3
python-dotenv==1.0.0
numpy==1.23.5
pandas==1.5.3
sqlalchemy==2.0.9
cx-oracle==8.3.0
vertica-python==1.1.1
aiohttp==3.8.4
python-multipart==0.0.6
requests==2.28.2
apscheduler==3.10.1
psycopg2-binary==2.9.6
pytest
pytest-cov
google-cloud-secret-manager
mistralai 